+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2025-07-04 - Feature: Show course progress as progress bar on the course cards, resolves #844
 * 2025-07-03 - Improvement: Show the login link in the navbar as button, resolves #979.
 * 2025-06-06 - Feature: Add color picker settings for calendar branding, resolves #434.
 * 2025-06-30 - Regression: Flavour favicon images were not working, resolves #942